+++ debug_utils/sierra-emu/src/vm/int.rs
@@ -7,6 +7,7 @@ use cairo_lang_sierra::{
             signed::{SintConcrete, SintTraits},
             signed128::Sint128Concrete,
             unsigned::{UintConcrete, UintTraits},
+            unsigned128::Uint128Concrete,
             IntConstConcreteLibfunc, IntMulTraits, IntOperationConcreteLibfunc, IntOperator,
             IntTraits,
         },
@@ -15,7 +16,7 @@ use cairo_lang_sierra::{
     },
     program_registry::ProgramRegistry,
 };
-use num_bigint::BigInt;
+use num_bigint::{BigInt, BigUint};
 use num_traits::ops::overflowing::{OverflowingAdd, OverflowingSub};
 use smallvec::smallvec;
 use starknet_crypto::Felt;
@@ -27,13 +28,13 @@ use crate::{
 
 use super::EvalAction;
 
-fn apply_wrapping_op(
+fn apply_operation_for_type(
     ty: &CoreTypeConcrete,
     lhs: BigInt,
     rhs: BigInt,
     op: IntOperator,
-) -> (BigInt, usize) {
-    fn wrapping<T>(lhs: BigInt, rhs: BigInt, op: IntOperator) -> (BigInt, usize)
+) -> (BigInt, bool) {
+    fn operation<T>(lhs: BigInt, rhs: BigInt, op: IntOperator) -> (BigInt, bool)
     where
         T: OverflowingAdd + OverflowingSub + Into<BigInt> + TryFrom<BigInt>,
         <T as TryFrom<BigInt>>::Error: Debug,
@@ -41,26 +42,26 @@ fn apply_wrapping_op(
         let lhs: T = lhs.try_into().unwrap();
         let rhs: T = rhs.try_into().unwrap();
 
-        let (res, ovf) = match op {
+        let (result, had_overflow) = match op {
             IntOperator::OverflowingAdd => OverflowingAdd::overflowing_add(&lhs, &rhs),
             IntOperator::OverflowingSub => OverflowingSub::overflowing_sub(&lhs, &rhs),
         };
 
-        (res.into(), ovf as usize)
+        (result.into(), had_overflow)
     }
 
     match ty {
-        CoreTypeConcrete::Sint8(_) => wrapping::<u8>(lhs, rhs, op),
-        CoreTypeConcrete::Sint16(_) => wrapping::<i16>(lhs, rhs, op),
-        CoreTypeConcrete::Sint32(_) => wrapping::<i32>(lhs, rhs, op),
-        CoreTypeConcrete::Sint64(_) => wrapping::<i64>(lhs, rhs, op),
-        CoreTypeConcrete::Sint128(_) => wrapping::<i128>(lhs, rhs, op),
-        CoreTypeConcrete::Uint8(_) => wrapping::<u8>(lhs, rhs, op),
-        CoreTypeConcrete::Uint16(_) => wrapping::<u16>(lhs, rhs, op),
-        CoreTypeConcrete::Uint32(_) => wrapping::<u32>(lhs, rhs, op),
-        CoreTypeConcrete::Uint64(_) => wrapping::<u64>(lhs, rhs, op),
-        CoreTypeConcrete::Uint128(_) => wrapping::<u128>(lhs, rhs, op),
-        _ => panic!("Found a non-numeric type"),
+        CoreTypeConcrete::Sint8(_) => operation::<u8>(lhs, rhs, op),
+        CoreTypeConcrete::Sint16(_) => operation::<i16>(lhs, rhs, op),
+        CoreTypeConcrete::Sint32(_) => operation::<i32>(lhs, rhs, op),
+        CoreTypeConcrete::Sint64(_) => operation::<i64>(lhs, rhs, op),
+        CoreTypeConcrete::Sint128(_) => operation::<i128>(lhs, rhs, op),
+        CoreTypeConcrete::Uint8(_) => operation::<u8>(lhs, rhs, op),
+        CoreTypeConcrete::Uint16(_) => operation::<u16>(lhs, rhs, op),
+        CoreTypeConcrete::Uint32(_) => operation::<u32>(lhs, rhs, op),
+        CoreTypeConcrete::Uint64(_) => operation::<u64>(lhs, rhs, op),
+        CoreTypeConcrete::Uint128(_) => operation::<u128>(lhs, rhs, op),
+        _ => panic!("cannot apply integer operation to non-integer type"),
     }
 }
 
@@ -120,6 +121,27 @@ where
     }
 }
 
+pub fn eval_uint128(
+    registry: &ProgramRegistry<CoreType, CoreLibfunc>,
+    selector: &Uint128Concrete,
+    args: Vec<Value>,
+) -> EvalAction {
+    match selector {
+        Uint128Concrete::Const(info) => eval_const(registry, info, args),
+        Uint128Concrete::Operation(info) => eval_operation(registry, info, args),
+        Uint128Concrete::SquareRoot(info) => eval_square_root(registry, info, args),
+        Uint128Concrete::Equal(info) => eval_equal(registry, info, args),
+        Uint128Concrete::ToFelt252(info) => eval_to_felt(registry, info, args),
+        Uint128Concrete::FromFelt252(info) => eval_from_felt(registry, info, args),
+        Uint128Concrete::IsZero(info) => eval_is_zero(registry, info, args),
+        Uint128Concrete::Divmod(info) => eval_divmod(registry, info, args),
+        Uint128Concrete::Bitwise(info) => eval_bitwise(registry, info, args),
+        Uint128Concrete::GuaranteeMul(info) => eval_guarantee_mul(registry, info, args),
+        Uint128Concrete::MulGuaranteeVerify(info) => eval_guarantee_verify(registry, info, args),
+        Uint128Concrete::ByteReverse(info) => eval_byte_reverse(registry, info, args),
+    }
+}
+
 fn eval_const<T: IntTraits>(
     registry: &ProgramRegistry<CoreType, CoreLibfunc>,
     info: &IntConstConcreteLibfunc<T>,
@@ -171,7 +193,7 @@ fn eval_diff(
         .unwrap();
 
     let overflow = (lhs >= rhs) as usize;
-    let (res, _) = apply_wrapping_op(int_ty, lhs, rhs, IntOperator::OverflowingSub);
+    let (res, _) = apply_operation_for_type(int_ty, lhs, rhs, IntOperator::OverflowingSub);
     let res = get_value_from_integer(registry, int_ty, res);
 
     EvalAction::NormalBranch(overflow, smallvec![range_check, res])
@@ -270,10 +292,10 @@ fn eval_operation(
         .get_type(&info.signature.branch_signatures[0].vars[1].ty)
         .unwrap();
 
-    let (res, overflow) = apply_wrapping_op(int_ty, lhs, rhs, info.operator);
+    let (res, had_overflow) = apply_operation_for_type(int_ty, lhs, rhs, info.operator);
     let res = get_value_from_integer(registry, int_ty, res);
 
-    EvalAction::NormalBranch(overflow, smallvec![range_check, res])
+    EvalAction::NormalBranch(had_overflow as usize, smallvec![range_check, res])
 }
 
 fn eval_square_root(
@@ -322,4 +344,48 @@ fn eval_widemul(
     let res = lhs * rhs;
 
     EvalAction::NormalBranch(0, smallvec![get_value_from_integer(registry, int_ty, res)])
+}
+
+fn eval_guarantee_mul(
+    _registry: &ProgramRegistry<CoreType, CoreLibfunc>,
+    _info: &SignatureOnlyConcreteLibfunc,
+    args: Vec<Value>,
+) -> EvalAction {
+    let [Value::U128(lhs), Value::U128(rhs)]: [Value; 2] = args.try_into().unwrap() else {
+        panic!()
+    };
+
+    let mask128 = BigUint::from(u128::MAX);
+    let result = BigUint::from(lhs) * BigUint::from(rhs);
+    let high = Value::U128((&result >> 128u32).try_into().unwrap());
+    let low = Value::U128((result & mask128).try_into().unwrap());
+
+    EvalAction::NormalBranch(0, smallvec![high, low, Value::Unit])
+}
+
+fn eval_guarantee_verify(
+    _registry: &ProgramRegistry<CoreType, CoreLibfunc>,
+    _info: &SignatureOnlyConcreteLibfunc,
+    args: Vec<Value>,
+) -> EvalAction {
+    let [range_check @ Value::Unit, _verify @ Value::Unit]: [Value; 2] = args.try_into().unwrap()
+    else {
+        panic!()
+    };
+
+    EvalAction::NormalBranch(0, smallvec![range_check])
+}
+
+fn eval_byte_reverse(
+    _registry: &ProgramRegistry<CoreType, CoreLibfunc>,
+    _info: &SignatureOnlyConcreteLibfunc,
+    args: Vec<Value>,
+) -> EvalAction {
+    let [bitwise @ Value::Unit, Value::U128(value)]: [Value; 2] = args.try_into().unwrap() else {
+        panic!()
+    };
+
+    let value = value.swap_bytes();
+
+    EvalAction::NormalBranch(0, smallvec![bitwise, Value::U128(value)])
 }
